Make a different pyproject.toml file for Playwright tests running under 3.13. Put tests in their own dir.
