+++ .github/workflows/playwright.yml
@@ -4,27 +4,18 @@ on: [ push, pull_request ]
 jobs:
   playwright:
     runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [ '3.13' ]
-    name: Python ${{ matrix.python-version }}
+    name: Setup Playwright
     steps:
       - uses: actions/checkout@v4
       - name: Install the latest version of uv
         uses: astral-sh/setup-uv@v6
         with:
           version: "latest"
-      - name: Install Python 3.13
-        if: matrix.python-version == '3.13'
+      - name: Setup environment for Python 3.13
         run: |
-          sudo add-apt-repository ppa:deadsnakes/nightly
-          sudo apt-get update
-          sudo apt-get install -y --no-install-recommends python3.13-dev python3.13-venv
-          python3.13 -m pip install --upgrade pip setuptools
-          python3.13 -m venv $HOME/venv-python3.13
+          cp pyproject-playwright.toml pyproject.toml
           uv sync
-          uv add --dev pytest-playwright
       - name: Run tests
         run: |
           uv run playwright install
-          uv run pytest -v -m "playwright"
+          uv run pytest
