+++ tests/conftest.py
@@ -1,1 +1,6 @@
-pytest_plugins = "tdom.fixtures"
+try:
+    import playwright
+
+    pytest_plugins = "tdom.fixtures"
+except ImportError:
+    pass
