+++ tests/pwright/stubs/hello/index.html
@@ -1,21 +1,3 @@
-<!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>tdom - test runner</title>
-    <link rel="stylesheet" href="https://pyscript.net/releases/2025.3.1/core.css">
-    <script type="module" src="https://pyscript.net/releases/2025.3.1/core.js"></script>
-    <script type="py" src="./index.py" config="./index.toml"></script>
-</head>
-<body>
-<div id="output"></div>
-<py-script>
-from pyscript import document
-from index import main
-
-result = main()
-document.querySelector("#output").innerHTML = result
-</py-script>
-</body>
+<head><title>tdom</title></head>
 </html>
