--- tests/stubs/hello/index.html
@@ -1,3 +1,0 @@
-<html lang="en">
-<head><title>tdom</title></head>
-</html>
